change priorityLevel to enum (1-5)

+++ kernel/API/ProcessCtl.h
@@ -73,7 +73,7 @@ typedef struct ProcessInfo
     Process::State state;
 
     /** Defines the current priority of the Process. */
-    uint priority;
+    u8 priority;
 }
 ProcessInfo;
